Updated createTestImage() with an additional argument that allows a user to specify the returned datatype of the test image.

+++ gaip/tests/unittesting_tools.py
@@ -76,7 +76,7 @@ class ParameterisedTestCase(unittest.TestCase):
 
 
 def createTestImage(dimensions=(1000,1000), geotransform=None,
-        projection=None, resolution=(25.0,25.0)):
+        projection=None, resolution=(25.0,25.0), dtype='uint8'):
     """
     Creates an image with geo-location information.
 
@@ -118,7 +118,7 @@ def createTestImage(dimensions=(1000,1000), geotransform=None,
         dimensions (y, x), containing values in the range [0,256).
         The 2nd element contains an instance of a GriddedGeoBox.
     """
-    img = numpy.random.randint(0, 256, dimensions).astype('uint8')
+    img = numpy.random.randint(0, 256, dimensions).astype(dtype)
 
     if (geotransform is None) or (projection is None):
         geotransform = (635000.0, 25.0, 0.0, 6277000.0, 0.0, 25.0)
